refactor: split filesystem related code into new handler struct
... so the actual filesystem server is just the regular `server.MCPServer` struct.
This way it can be used with the in-process client.

Also added a test to cover this functionality.
